remove integration test to ConsoleConsumerTest
